Expression system and minor refactoring: * NEW: Expression system * ADD: (UI) HintWindow expression support * ADD: (UI) ProgressBar expression support * ADD: xr_string is now our class, not just typedef * EDIT: CLevel::net_start_client2 no anymore update client and server * EDIT: CScriptBinder::net_Spawn remove 'try catch' * EDIT: g_pStringContainer is now static variable

+++ code/engine.vc2008/xrCore/_stl_extensions.h
@@ -163,7 +163,7 @@ public:
 
     bool StartWith(const xr_string& Other) const;
     bool StartWith(LPCSTR Str) const;
-    bool StartWith(LPCSTR Str, int Size) const;
+    bool StartWith(LPCSTR Str, size_t Size) const;
     xr_string RemoveWhitespaces() const;
 };
